pyproject.toml
[tool.poetry]
name = "gas-goblin"
version = "0.1.0"
description = ""
authors = ["tobinio <Tobias.frischmann1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uagents = "^0.15.2"
uagents-ai-engine = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
